XML, 예외처리 분리완료

+++ backend/src/main/resources/sql/schema.sql
@@ -3,11 +3,10 @@ DROP TABLE IF EXISTS purchases;
 DROP TABLE IF EXISTS goods_price;
 DROP TABLE IF EXISTS goods_basic;
 
-
 -- 2. 공매 기본 정보
 CREATE TABLE goods_basic (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT COMMENT '기본 ID',
-    history_no BIGINT UNIQUE NOT NULL COMMENT '물건이력번호',
+    id INT PRIMARY KEY AUTO_INCREMENT COMMENT '기본 ID',
+    history_no INT UNIQUE NOT NULL COMMENT '물건이력번호',
     goods_name VARCHAR(1000) NOT NULL COMMENT '물건명',
     status_name VARCHAR(100) COMMENT '물건상태',
     sale_type_name VARCHAR(100) COMMENT '처분방식명',
@@ -23,8 +22,8 @@ CREATE TABLE goods_basic (
 
 -- 3. 공매 가격 정보
 CREATE TABLE goods_price (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT COMMENT '가격 ID',
-    history_no BIGINT NOT NULL UNIQUE COMMENT '물건이력번호',
+    id INT PRIMARY KEY AUTO_INCREMENT COMMENT '가격 ID',
+    history_no INT NOT NULL UNIQUE COMMENT '물건이력번호',
     min_bid_price BIGINT NOT NULL COMMENT '최저입찰가',
     appraisal_price BIGINT COMMENT '감정가',
     fee_rate VARCHAR(20) COMMENT '최저입찰가율',
@@ -36,8 +35,8 @@ CREATE TABLE goods_price (
 
 -- 4. 구매 이력
 CREATE TABLE purchases (
-    id BIGINT PRIMARY KEY AUTO_INCREMENT COMMENT '매매 ID',
-    history_no BIGINT NOT NULL COMMENT '물건이력번호',
+    id INT PRIMARY KEY AUTO_INCREMENT COMMENT '매매 ID',
+    history_no INT NOT NULL COMMENT '물건이력번호',
     purchase_price BIGINT NOT NULL COMMENT '구매가격',
     purchase_status VARCHAR(20) DEFAULT 'COMPLETED' COMMENT '구매상태 (PENDING, COMPLETED, CANCELLED)',
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP COMMENT '구매일시',
